+++ packages/moonstone/Slider/Slider.js
@@ -1,7 +1,10 @@
 /**
- * Exports the {@link moonstone/Slider.Slider} component.
+ * Provides a Moonstone-themed slider components and behaviors
  *
  * @module moonstone/Slider
+ * @exports Slider
+ * @exports SliderBase
+ * @exports SliderDecorator
  */
 
 import {forKey, forProp, forward, forwardWithPrevent, handle} from '@enact/core/handle';
@@ -31,8 +34,8 @@ import componentCss from './Slider.less';
  * Range-selection input component
  *
  * @class Slider
+ * @extends ui/Slider.Slider
  * @memberof moonstone/Slider
- * @mixes moonstone/Slider.SliderDecorator
  * @ui
  * @public
  */
@@ -59,77 +62,25 @@ const SliderBase = kind({
 		active: PropTypes.bool,
 
 		/**
-		 * Background progress, as a proportion between `0` and `1`.
-		 *
-		 * @type {Number}
-		 * @default 0
-		 * @public
-		 */
-		backgroundProgress: PropTypes.number,
-
-		/**
 		 * Customizes the component by mapping the supplied collection of CSS class names to the
 		 * corresponding internal Elements and states of this component.
 		 *
+		 * The following classes are supported:
+		 *
+		 * * `slider` - The root component class
+		 *
 		 * @type {Object}
-		 * @private
+		 * @public
 		 */
 		css: PropTypes.object,
 
 		/**
-		 * When `true`, the component is shown as disabled and does not generate events
-		 *
-		 * @type {Boolean}
-		 * @public
-		 */
-		disabled: PropTypes.bool,
-
-		/**
 		 * When `true`, the tooltip, if present, is shown
+		 *
 		 * @type {Boolean}
 		 * @public
 		 */
 		focused: PropTypes.bool,
-		knob: PropTypes.node,
-
-		/**
-		 * The amount to increment or decrement the position of the knob via 5-way controls.
-		 * When `detachedKnob` is false, the knob must first be activated by selecting it. When
-		 * `detachedKnob` is true, the knob will respond to direction key presses without
-		 * activation.
-		 *
-		 * If not specified, `step` is used for the default value.
-		 *
-		 * @type {Number}
-		 * @public
-		 */
-		knobStep: PropTypes.number,
-
-		/**
-		 * The maximum value of the slider.
-		 *
-		 * @type {Number}
-		 * @default 100
-		 * @public
-		 */
-		max: PropTypes.number,
-
-		/**
-		 * The minimum value of the slider.
-		 *
-		 * @type {Number}
-		 * @default 0
-		 * @public
-		 */
-		min: PropTypes.number,
-
-		/**
-		 * When `true`, the slider bar doesn't show a fill and doesn't highlight when spotted
-		 *
-		 * @type {Boolean}
-		 * @public
-		 */
-		noFill: PropTypes.bool,
 
 		/**
 		 * The handler when the knob is activated or deactivated by selecting it via 5-way
@@ -140,58 +91,13 @@ const SliderBase = kind({
 		onActivate: PropTypes.func,
 
 		/**
-		 * The handler to run when the value is changed.
-		 *
-		 * @type {Function}
-		 * @param {Object} event
-		 * @param {Number} event.value Value of the slider
-		 * @public
-		 */
-		onChange: PropTypes.func,
-
-		/**
-		 * If `true` the slider will be oriented vertically.
-		 *
-		 * @type {Boolean}
-		 * @default false
-		 * @public
-		 */
-		orientation: PropTypes.string,
-
-		/**
-		 * When `true`, a pressed visual effect is applied
-		 *
-		 * @type {Boolean}
-		 * @public
-		 */
-		pressed: PropTypes.bool,
-
-		/**
-		 * The amount to increment or decrement the value.
-		 *
-		 * @type {Number}
-		 * @default 1
-		 * @public
-		 */
-		step: PropTypes.number,
-
-		/**
 		 * Enables the built-in tooltip, whose behavior can be modified by the other tooltip
 		 * properties.
 		 *
 		 * @type {Boolean|Element|Function}
 		 * @public
 		 */
-		tooltip: PropTypes.oneOfType([PropTypes.bool, PropTypes.object, PropTypes.func]),
-
-		/**
-		 * The value of the slider.
-		 *
-		 * @type {Number}
-		 * @default 0
-		 * @public
-		 */
-		value: PropTypes.number
+		tooltip: PropTypes.oneOfType([PropTypes.bool, PropTypes.object, PropTypes.func])
 	},
 
 	defaultProps: {
@@ -247,18 +153,17 @@ const SliderBase = kind({
 		}
 	},
 
-	render: ({css, knob, ...rest}) => {
-		delete rest.activateOnFocus;
-		delete rest.active;
-		delete rest.focused;
-		delete rest.onActivate;
-		delete rest.tooltip;
+	render: (props) => {
+		delete props.activateOnFocus;
+		delete props.active;
+		delete props.focused;
+		delete props.onActivate;
+		delete props.tooltip;
 
 		return (
 			<UiSlider
-				{...rest}
-				css={css}
-				knobComponent={knob}
+				{...props}
+				css={props.css}
 			/>
 		);
 	}
