+++ packages/ui/Slider/Slider.less
@@ -17,6 +17,7 @@
 	}
 
 	.knob {
+		--ui-slider-knob-proportion: 0;
 		top: 50%;
 		left: 50%;
 	}
